python/requirements.txt
-e ../agent-s_repo
pyautogui
pillow
requests
flask
airllm
pytesseract
